Fixed the problems like changing from BCI to Neuro,Reformatting code, updating gradle dependencies and using string resources.
