+++ app/src/main/java/io/neurolab/MainActivity.java
@@ -70,7 +70,7 @@ public class MainActivity extends AppCompatActivity
 
         //noinspection SimplifiableIfStatement
         if (id == R.id.action_settings) {
-            startActivity(new Intent(MainActivity.this, BCISettingsActivity.class));
+            startActivity(new Intent(MainActivity.this, NeuroSettingsActivity.class));
             return true;
         }
