+++ app/src/main/java/io/neurolab/NeuroSettingsActivity.java
@@ -3,11 +3,12 @@ package io.neurolab;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class BCISettingsActivity extends AppCompatActivity {
+public class NeuroSettingsActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_bcisettings);
+        setContentView(R.layout.activity_neurosettings);
+        setTitle("Neuro Settings");
     }
 }
