+++ app/src/main/res/layout/activity_neurosettings.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:name="io.neurolab.BCISettingsFragment"
     xmlns:tools="http://schemas.android.com/tools"
+    android:name="io.neurolab.NeuroSettingsFragment"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".BCISettingsActivity">
+    tools:context=".NeuroSettingsActivity">
 
 </fragment>
