+++ app/src/main/res/values/strings.xml
@@ -3,6 +3,13 @@
 
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
-
     <string name="action_settings">Settings</string>
+
+    <string name="samples_pref_key">samples</string>
+    <string name="bins_pref_key">bins</string>
+    <string name="num_channels_pref_key">numChannels</string>
+
+    <string name="samples_pref_label">Samples per second</string>
+    <string name="bins_pref_label">Bins</string>
+    <string name="num_channels_pref_label">Number Channels</string>
 </resources>
