--- app/src/main/res/xml/bci_settings_fragment.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <EditTextPreference
-        android:defaultValue="1"
-        android:title="Samples per second"
-        android:key="samples"
-        />
-    <EditTextPreference
-        android:defaultValue="1"
-        android:title="Bins"
-        android:key="bins"
-        />
-    <EditTextPreference
-        android:defaultValue="1"
-        android:title="Num Channels"
-        android:key="numChannels"
-        />
-    <EditTextPreference
-        android:defaultValue="1"
-        android:title="Current Data"
-        android:key="current_data"
-        />
-</PreferenceScreen>
